changed variable types to float64

+++ wavenet/model.py
@@ -7,7 +7,7 @@ from .ops import causal_conv, mu_law_encode
 def create_variable(name, shape):
     """Create a convolution filter variable with the specified name and shape,
     and initialize it using Xavier initialition."""
-    initializer = tf.contrib.layers.xavier_initializer_conv2d()
+    initializer = tf.contrib.layers.xavier_initializer_conv2d(dtype=tf.float64)
     variable = tf.Variable(initializer(shape=shape), name=name)
     return variable
 
@@ -15,7 +15,7 @@ def create_variable(name, shape):
 def create_embedding_table(name, shape):
     if shape[0] == shape[1]:
         # Make a one-hot encoding as the initial value.
-        initial_val = np.identity(n=shape[0], dtype=np.float32)
+        initial_val = np.identity(n=shape[0], dtype=np.float64)
         return tf.Variable(initial_val, name=name)
     else:
         return create_variable(name, shape)
